Move Client and Server into their own modules.
This avoids the httplib2 dependency unless you need the client.
